Move postgres SSL test to external
